+++ src/client/src/pages/detail/form/stratigraphy/lithology/lithologyInfo/infoList/InfoList.jsx
@@ -1,105 +1,141 @@
-import { useCallback, useMemo, useState } from "react";
-import { Form, Input } from "semantic-ui-react";
-import _ from "lodash";
-import DateField from "../../../../../../../components/legacyComponents/dateField.jsx";
-import DomainDropdown from "../../../../../../../components/legacyComponents/domain/dropdown/domainDropdown.jsx";
-import TranslationText from "../../../../../../../components/legacyComponents/translationText.jsx";
-import * as Styled from "./styles.js";
+import { useContext, useEffect } from "react";
+import { Controller, FormProvider, useForm } from "react-hook-form";
+import { useTranslation } from "react-i18next";
+import { Box, FormControlLabel, Stack, Switch } from "@mui/material";
+import { Trash2 } from "lucide-react";
+import { DevTool } from "../../../../../../../../hookformDevtools";
+import { copyStratigraphy, deleteStratigraphy } from "../../../../../../../api/fetchApiV2";
+import { CancelButton, CopyButton, DeleteButton, SaveButton } from "../../../../../../../components/buttons/buttons";
+import { DataCardButtonContainer } from "../../../../../../../components/dataCard/dataCard";
+import { FormContainer, FormDomainSelect, FormInput, FormValueType } from "../../../../../../../components/form/form";
+import { ensureDatetime } from "../../../../../../../components/form/formUtils";
+import { PromptContext } from "../../../../../../../components/prompt/promptContext";
+import { DetailContext } from "../../../../../detailContext";
+import { updateStratigraphyAttributes } from "../api/index.ts";
 
-const InfoList = props => {
-  const { attribute, profileInfo, updateChange, isEditable } = props.data;
-  const [inputDisplayValues, setInputDisplayValues] = useState({});
+const InfoList = ({ id, profileInfo, onUpdated }) => {
+  const { t } = useTranslation();
+  const formMethods = useForm({ mode: "all" });
+  const { showPrompt } = useContext(PromptContext);
+  const { editingEnabled } = useContext(DetailContext);
 
-  // This adds a delay to each keystroke before calling the updateChange method in the parent component.
-  // It prevents the API from being called too often.
-  const debouncedUpdateChange = useMemo(
-    () =>
-      _.debounce((value, newValue, to, isNumber) => {
-        updateChange(value, newValue, to, isNumber);
-      }, 2000),
-    [updateChange],
-  );
+  useEffect(() => {
+    formMethods.reset({
+      name: profileInfo.name,
+      date: profileInfo.date?.toString().slice(0, 10) ?? null,
+      qualityId: profileInfo.qualityId,
+      isPrimary: profileInfo.isPrimary,
+    });
+  }, [profileInfo, formMethods]);
 
-  const updateInputDisplayValue = useCallback(
-    (value, inputValue) => {
-      setInputDisplayValues(prevInputDisplayValues => ({
-        ...prevInputDisplayValues,
-        [value]: inputValue,
-      }));
-    },
-    [setInputDisplayValues],
-  );
+  const submitForm = data => {
+    data.date = data?.date ? ensureDatetime(data.date.toString()) : null;
+    updateStratigraphyAttributes(id, data);
+  };
+
+  useEffect(() => {
+    formMethods.setValue("date", profileInfo?.date?.toString().slice(0, 10) ?? "");
+  }, [profileInfo?.date, formMethods]);
 
   return (
-    <>
-      <Styled.FormContainer>
-        {attribute?.map((item, key) => (
-          <Form autoComplete="false" error key={key} size="small">
-            <Styled.AttributesContainer required={item.require}>
-              <Styled.Label>
-                <TranslationText id={item.label} />
-              </Styled.Label>
-
-              {item.type === "Input" && (
-                <Styled.AttributesItem>
-                  <Input
-                    data-cy={item.value}
-                    autoCapitalize="off"
-                    autoComplete="off"
-                    autoCorrect="off"
-                    readOnly={!isEditable}
-                    onChange={e => {
-                      updateInputDisplayValue(item.value, e.target.value);
-                      debouncedUpdateChange(
-                        item.value,
-                        e.target.value === "" ? null : e.target.value,
-                        item?.to,
-                        item?.isNumber,
-                      );
-                    }}
-                    spellCheck="false"
-                    style={{ width: "100%" }}
-                    value={
-                      _.isNil(inputDisplayValues[item.value])
-                        ? _.isNil(profileInfo?.[item.value])
-                          ? ""
-                          : profileInfo[item.value]
-                        : inputDisplayValues[item.value]
-                    }
+    <FormProvider {...formMethods}>
+      <DevTool control={formMethods.control} placement="top-left" />
+      <FormContainer>
+        <FormContainer direction={"row"}>
+          <FormInput
+            fieldName={"name"}
+            label={"stratigraphy_name"}
+            value={profileInfo.name}
+            readonly={!editingEnabled}
+            type={FormValueType.Text}></FormInput>
+          <FormControlLabel
+            control={
+              <Controller
+                name="isPrimary"
+                control={formMethods.control}
+                defaultValue={profileInfo?.isPrimary}
+                render={({ field }) => (
+                  <Switch
+                    {...field}
+                    data-cy={"isprimary-switch"}
+                    checked={field.value}
+                    color="secondary"
+                    disabled={!editingEnabled || profileInfo?.isPrimary}
                   />
-                </Styled.AttributesItem>
-              )}
-              {item.type === "Dropdown" && (
-                <Styled.AttributesItem>
-                  <DomainDropdown
-                    data-cy={item.value}
-                    multiple={item.multiple}
-                    readOnly={!isEditable}
-                    onSelected={e => updateChange(item.value, item.multiple ? e.map(mlpr => mlpr.id) : e.id, false)}
-                    schema={item.schema}
-                    search={item.search}
-                    selected={_.isNil(profileInfo?.[item.value]) ? null : profileInfo[item.value]}
-                  />
-                </Styled.AttributesItem>
-              )}
-
-              {item.type === "Date" && (
-                <Styled.AttributesItem>
-                  <DateField
-                    isEditable={isEditable}
-                    data-cy={item.value}
-                    date={profileInfo?.[item.value] ? profileInfo[item.value] : null}
-                    onChange={selected => {
-                      updateChange(item.value, selected, false);
-                    }}
-                  />
-                </Styled.AttributesItem>
-              )}
-            </Styled.AttributesContainer>
-          </Form>
-        ))}
-      </Styled.FormContainer>
-    </>
+                )}
+              />
+            }
+            label={t("mainStratigraphy")}
+          />
+        </FormContainer>
+        <FormContainer direction={"row"}>
+          <FormInput
+            fieldName={"date"}
+            label="date"
+            readonly={!editingEnabled}
+            value={profileInfo.date}
+            type={FormValueType.Date}
+          />
+          <FormDomainSelect
+            fieldName={"qualityId"}
+            label={"stratigraphy_quality"}
+            selected={profileInfo.qualityId}
+            readonly={!editingEnabled}
+            schemaName={"description_quality"}
+          />
+        </FormContainer>
+        <Stack direction={"row"} gap={1}>
+          <Box sx={{ flexGrow: 1 }} />
+          {editingEnabled && (
+            <>
+              <DataCardButtonContainer>
+                <CopyButton
+                  onClick={() => {
+                    copyStratigraphy(profileInfo?.id).then(() => {
+                      onUpdated("cloneStratigraphy");
+                    });
+                  }}></CopyButton>
+                <DeleteButton
+                  onClick={() => {
+                    showPrompt(t("deleteMessage"), [
+                      {
+                        label: t("cancel"),
+                        action: null,
+                      },
+                      {
+                        label: t("delete"),
+                        icon: <Trash2 />,
+                        variant: "contained",
+                        action: () => {
+                          deleteStratigraphy(profileInfo?.id).then(() => {
+                            onUpdated("deleteStratigraphy");
+                          });
+                        },
+                      },
+                    ]);
+                  }}
+                />
+              </DataCardButtonContainer>
+              <DataCardButtonContainer>
+                <CancelButton
+                  dataCy={"stratigraphy-cancel-button"}
+                  onClick={() => {
+                    formMethods.reset();
+                  }}
+                />
+                <SaveButton
+                  dataCy={"" + "stratigraphy-save-button"}
+                  disabled={!formMethods.formState.isValid}
+                  onClick={() => {
+                    formMethods.handleSubmit(submitForm)();
+                  }}
+                />
+              </DataCardButtonContainer>
+            </>
+          )}
+        </Stack>
+      </FormContainer>
+    </FormProvider>
   );
 };